Purpose: Bug fix
Description:
This checkin addresses two bugs.
Firstly, I used the wrong semantics when detecting if $MAKE is defined to
the empty string or undefined.
Secondly, pmake on IRIX does not work well with automake's dependency tracking
if the build is not an in-place build.

Solution:
Fixed semantics in configure.in.
Disabled automatic dependency tracking automatically on IRIX.  This is not
the optimal fix, but it should work fine as long as no one is doing HDF5
development on IRIX, and it is much easier than trying to detect the exact
situations in which dependency tracking fails (or than trying to fix dependency
tracking!).

Platforms tested:
heping, modi4 (pmake and gmake)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10064 dab4d1a6-ed17-0410-a064-d1ae371a2980