Purpose:  Bug fix
Description:  On some Cray, size of "short" is 8 byte but only the low-order 4 byte
are ever used(precision 4 byte).  The "perm" in struct "detect_t" contains `-1' for
bytes that aren't used, making detection of order invalid in function print_results().

Solution: Use a "for" loop to search through "perm" find the first non "-1" element,
and use it to decide the byte order of the machine.

Platforms tested:  Cray SV1 and fuss - limited change.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10069 dab4d1a6-ed17-0410-a064-d1ae371a2980