Purpose: Bug fix
Description:
Libtool 1.5.14 added some extra flags that needed to be passed from the compiler
to the linker.  icc did not handle this correctly and complained when it saw
unknown flags.

Solution:
Set the -Wl flag correctly on freebsd so that icc knows to pass flags through
to the linker.

Platforms tested:
sleipnir (gcc, g++, icc, gcc40)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10106 dab4d1a6-ed17-0410-a064-d1ae371a2980