Purpose:  Bug fix
Description:  In H5T_conv_f_f(), for the special case of +/-infinity,
if datatype has no implied mantissa bit, the 1st bit of mantissa should
be set to 1 to make an infinity(all other mantissa bits should be 0 and
all exponent bits should be 1).  The Intel-Linux "long double" is this
case.

Solution:  Added this case during the check for special cases.

Platforms tested: fuss - tested with h5committest for v1.7


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10405 dab4d1a6-ed17-0410-a064-d1ae371a2980