Purpose: Feature
Description:
Added new function, getenv_all (only for parallel builds)

Solution:
getenv_all is a collective version of getenv.  It is used to the return the value of
an environment variable in another task and can be used to synchronize all of the
environment variables used by all of the tasks.  This helps to fix problems that commonly crop up
(like with HDF5_PARAPREFIX) when an environment variable does not propogate to other tasks.

This commit only includes the function.  Future commits will replace the various getenv calls with getenv_all calls.

Platforms tested:
heping, sol, copper

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10555 dab4d1a6-ed17-0410-a064-d1ae371a2980