Purpose: Bug fix.
Description:
When there were errors in the test, program still just call MPI_FINALIZE
and failed to attempt to exit with some error code so that calling
programs like make be informed of the exceptions.

Solution:
Call MPI_ABORT if error is detected. Though MPI_ABORT does not
guarantee 100% failure report, it has the best chance.
Also made dimension incompatible as a real error.

Platforms tested:
Did not h5committest but tested in heping PP only since this
is a simple fix.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10641 dab4d1a6-ed17-0410-a064-d1ae371a2980