Purpose: Code cleanup
Description:
On Cray X1, a loop in the Fortran C source confuses the compiler's optimizer
and hangs the build.

Solution:
Documented the fix in the code, and added an #ifdef so that
the optimization is only disabled on Cray X1.

Platforms tested:
mir (fix was already checked into code; this was a minor change)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10774 dab4d1a6-ed17-0410-a064-d1ae371a2980