Purpose: Bug fix (kind of...)
Description:
When trying to 'make distclean', the Makefiles will attempt to re-run
configure if configure has changed.  This is a nice idea, but prevents
the distclean from happening if configure fails (e.g., if the environment
has changed).  This happens frequently in the Daily Tests when configure
is altered.

Solution:
It turns out to be easier than I had feared to override the dependency
that is causing configure to be re-run.  Now configure will never
be run automatically, even if it is out-of-date.
It would be more complicated to cause it to be re-run in all cases except
distclean, so we'll try just the inelegant fix for now.

Platforms tested:
mir.  This change doesn't affect the source, only what is built when.
The real test will be how the Daily Tests handle changes to configure
(like the one I checked in earlier today).


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10775 dab4d1a6-ed17-0410-a064-d1ae371a2980