Purpose: Cleanup.
Description:
The parallel fortran test was entered under $TEST_PROGS which is for
serial tests.  This worked because $RUNTEST is hardset to
$RUNPARALLEL.

Solution:
Moved the parallel test to $TEST_PROGS_PARA and removed the
setting of $RUNTEST.

Platforms tested:
Tested in heping in PP mode only as this is pretty trivial.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10830 dab4d1a6-ed17-0410-a064-d1ae371a2980