Purpose: bug fix
Description:
check-s and check-p were set as pre-requistics of test.  They
would get executed in parallel if parallel make is used.  This
could cause problems since serial tests are used to be executed
before parallel tests. It is not known if it is always okay to
run serial and parallel tests in simultantously.

Solution:
Change check-s and check-p as actions of the target test so
that they get executed sequnentially.

Platforms tested:
heping PP.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10849 dab4d1a6-ed17-0410-a064-d1ae371a2980