Purpose: Bug fix
Description:
Fortran type detection failed on Cygwin.  This is because executables
in Cygwin have the .exe extension, which confused dependencies in
the Makefile.

Solution:
Sprinkled the $(EXEEXT) variable liberally about the fortran/src
Makefile.am.  This variable is set at configure time and will
be ".exe" on Cygwin and stay "" on Linux.

Platforms tested:
Cygwin, modi4, mir


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10870 dab4d1a6-ed17-0410-a064-d1ae371a2980