Purpose: Work Around A Compiler Bug
Description: PGCC compiler on mir(AMD Opeteron) has a bug for optimization,
causing the multi driver test to fail.  The compiler handles a macro in this
file incorrectly.

Solution: Slightly change the definition of this macro to avoid the problem
temporarily.

Platforms tested: h5committest, optimization and debug on mir.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10913 dab4d1a6-ed17-0410-a064-d1ae371a2980