Purpose: Bug fix.
Description: There was a bug in the calculating code of the Fletcher32
checksum in the library before v1.6.3.  The checksum value wasn't consistent
between big-endian and little-endian systems.  This bug was fixed in
Release 1.6.3.  However, after fixing the bug, the checksum value is no
longer the same as before on little-endian system.

Solution: Made the library compare both the correct checksum and incorrect
checksum generated from v1.6.2 or before.  This makes the library be
backward compatible but not forward compatible.

Platforms tested: h5committest and fuss.

Misc. update: Documented the forward incompatibility problem in the
Known Problem section in RELEASE.txt


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11054 dab4d1a6-ed17-0410-a064-d1ae371a2980