Purpose: Bug fix and improvement.
Description:
The test output were not displayed for parallel tests.
The serial tests output were always displayed whether the test was
actually ran this time or not.


Solution:
Moved the display of test output inside the target where the actual
tests are run.


Platforms tested:
Tested in heping using pp and fortran.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11213 dab4d1a6-ed17-0410-a064-d1ae371a2980