Purpose: bug fix
Description:

an assertion after the search for objects was saying that the number of
objects in file should be > 0
removed it, since there can be a case where the number of objects is 0.
in this case this was found by a case where zlib was not present and the output file
was not written any objects to

note: this was only detected on the windows batch files, that do not have a conditional
condition to not run tests when a filter is not present


Solution:

Platforms tested:
linux

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11584 dab4d1a6-ed17-0410-a064-d1ae371a2980