Purpose: Bug fix.
Description:
The previous way of detecting the MPICH version fails if the suspected
$CC does not give the expected MPICH version information.  Though no
consequence, it generates error messages that alarm users.
Also, mpich2 gives the version information as if it were mpich v1.0.2
which was incorrectly included as an earlier version of the origianl
mpich.

Solution:
Implemented a more robust code to detect the version and narrow the
recognition down to v1.2.0-v1.2.5.  Also, the code were duplicated in
two files (linux-gnulibc1 and ia64-linux-gnu).  Combined them into a
new file called mpich.

Platforms tested:
Tested in heping using mpicc and mpich2/mpicc and also in tg-ncsa.

Misc. update:
Updated MANIFEST.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11610 dab4d1a6-ed17-0410-a064-d1ae371a2980