Purpose: Patch.
Description:
Copper would fail with a message of
0032-113 Out of memory  in routine unknown, task 0
when run with 3 processes and size 1MB in MPI-IO tests.
It seems to be a copper MPIO error.

Solution:
Reduced the upper bound of default write size to 1/2MB (but
tests only go to 1/4MB) for now, pending permenant fix from
Copper.

Platforms tested:
Copper.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11788 dab4d1a6-ed17-0410-a064-d1ae371a2980