Purpose: Bug fix.
Description:
release created temporary files in .. directory.  If the -d directory
is the same as the parent directory, it would fail when it tried to
copy the temporary files to the destination directory.

Solution:
Changed release to create a temporary directory and create all temporary
files in that temporary directory.

Platforms tested:
Handtested in heping.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@11862 dab4d1a6-ed17-0410-a064-d1ae371a2980