Purpose: Removed -march flag
Description:
The -march flag was used for architecture-specific optimizations on a few
platforms.  However, it isn't always desirable.

Solution:
A user wanted to build HDF5 to be shared across several Linux
boxes, but the -march flag prevented it.

Platforms tested:
heping (one of the few platforms affected by this change)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12152 dab4d1a6-ed17-0410-a064-d1ae371a2980