Purpose: H5_CFLAGS change
Description:
Made the threadsafe flags in config/freebsd H5_CFLAGS so that they are
not passed on to h5cc.

Solution:
If user applications need the threadsafe flag to link properly against
the threadsafe HDF5 library, we'll see linking errors in make-install
when threadsafe is enabled, and can revert this change.

Platforms tested:
mir only (very minor change)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12312 dab4d1a6-ed17-0410-a064-d1ae371a2980