Purpose: Bug fix
Description:
The Daily Test errors were due to H5_CFLAGS not being included when
configure tested the sizes of some values.
Rolled back the offending changes.

Solution:
There is a "correct" solution to this problem, which is to ensure that
the H5_CFLAGS are included in autoconf's tests.
This rollback is only temporary until the "real" fix can be applied.

Platforms tested:
copper (problem only in aix config file)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12325 dab4d1a6-ed17-0410-a064-d1ae371a2980