Purpose: Upgrade
Description:
chnaged it from doing CVS-managed files to SVN-managed files.
(The use of ed, is real ugly, but ed exists in all Unix systems.
May want to consider using awk which is a more powerful and will have
a cleaner implementation.)

Platform tested:
Tested by hand in heping and shanti.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12485 dab4d1a6-ed17-0410-a064-d1ae371a2980