Fixed "make check-vfd"
"make check-vfd" will now run all tests in the test directory with different
file drivers (at least, all of those tests that use the testing framework's
FAPL).  Tests that fail will be skipped.

This is not a perfect fix, but is better than nothing.

Along with this change, check-vfd should be added to the Daily Tests.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12519 dab4d1a6-ed17-0410-a064-d1ae371a2980