H5L_link is called inside H5O.c but the header file that declared this function is not included in this file. Windows compiler issued a bad warning that"H5L_link" is not defined but is called.
So add header file "H5Lprivate.h" in H5O.c. Hopefully this is the right way.

Tested on windows and tested the compiling on Linux.
H5L_link is called inside H5O.c but the header file that declared this function is not included in this file. Windows compiler issued a bad warning that"H5L_link" is not defined but is called.

So add header file "H5Lprivate.h" in H5O.c. Hopefully this is the right way.

Tested on windows and tested the compiling on Linux.
H5L_link is called inside H5O.c but the header file that declared this function is not included in this file. Windows compiler issued a bad warning that"H5L_link" is not defined but is called.

So add header file "H5Lprivate.h" in H5O.c. Hopefully this is the right way.

Tested on windows and tested the compiling on Linux.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@12931 dab4d1a6-ed17-0410-a064-d1ae371a2980