Purpose: Bug fix.
Description:
When it reads the allhostfile, it does eliminate comments or blank lines
completely, thus it continues to use ssh/rsh even when only the local
host is tested.

Solution:
Changed the sed scripts to eliminate comments and blank lines.

Tested:
Cobalt by running bin/runtest.  Since this is not really library source,
the three platforms test does not apply.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13527 dab4d1a6-ed17-0410-a064-d1ae371a2980