Purpose: Bug problem.
Description:
CHunk allocation actually does not always work in parallel.  See Bug 281 and 636.
Case 2 test exposed the problem in an intermittent manner.  Skip the test by
default but allow it to be tested IF verbose is set to at least low.

Platform tested:
Kagiso parallel since that is where it sometimes failed.
The problem did not show up in copper.  The fix (skipping) is
straight forwarded.  So, no other machine tested.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13583 dab4d1a6-ed17-0410-a064-d1ae371a2980