Purpose: bug fix.
Description:
If a non-existing or empty source file is used with -c option, the
script ended up doing nothing and exit 0--a false success.  Changed
code so that it would do compile and end in error.

Also added an acknowledgement that the script is adapted from mpicc of MPICH.

Platform tested:
Kagiso (serial and PP).



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13600 dab4d1a6-ed17-0410-a064-d1ae371a2980