purpose: bug fix (760)
Description:
h5fc myprog.c libmylib.a does not work.  It complained missing routines that are in
libmylib.a.  The reason was because h5fc moved the libmylib.a to a separated libraries
variable but never put it back in the final linking stage.

Solution:
Removed the code that remove the *.a files.  Things work again though there is
sometimes a message complaining the *.a is not being used.  But the final a.out
is created nevertheless.  Need a more perfect solution.

Tested platforms:
kagiso.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13651 dab4d1a6-ed17-0410-a064-d1ae371a2980