h5repack and h5diff hyperslab I/O
changed the limit on which  hyperslab I/O is done from 1GB to 128MB

h5repack currently tests this feature by defining a dataset with dimensions of 128MB + 1byte (the datum being 1 byte integer), in which a 1Kb hyperslab was written. This  file is checked in SVN and its repacking is done by the program h5repacktst.c (along with the other h5repack tests). 

tested: linux



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13755 dab4d1a6-ed17-0410-a064-d1ae371a2980