Optimization of reading compound data.
Optimize a special case when the source members are a subset of
destination, and the order is the same, and no conversion is needed.
For example:
     struct source {            struct destination {
         TYPE1 A;      -->          TYPE1 A;
         TYPE2 B;      -->          TYPE2 B;
         TYPE3 C;      -->          TYPE3 C;
     };                             TYPE4 D;
                                    TYPE5 E;
                                };
The optimization is simply moving data to the appropriate
places in the buffer.  This optimization work is for the Chicago company.

Tested on sol, copper, smirom, liberty.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13803 dab4d1a6-ed17-0410-a064-d1ae371a2980