Purpose: Add c++ pubconf file.
Description:
Added static H5cxx_pubconf.h to our Windows zip file.  This file is manually maintained, as Windows does not run a configure script to generate it.  This file will also be added to the install directory when running the "installhdf5lib.bat" script.

Tested:
None, as the new macro isn't in use yet

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@13895 dab4d1a6-ed17-0410-a064-d1ae371a2980