Purpose: Fixes for 64-bit Windows Intel Fortran 9.1
Description:

We are porting Fortran to 64-bit Windows, and need to make many syntactical changes to bring the code up to date.  We now go back to using "!DEC$" rather than "!MS$" in our "ATTRIBUTE" code.  Also, the naming conventions for mixed language is inconsistent from 32- to 64- bit Windows.  To make up with this, we use the "decorate" flag.  We also needed to change the "calling convention" in project settings from "CVF" to "default".

Also other minor cleanups on project files.

Tested:
32-bit: VS.NET and VS2005
64-bit: VS2005

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14057 dab4d1a6-ed17-0410-a064-d1ae371a2980