Purpose: Revise usage of H5_HAVE_SIGLONGJMP
Description:
The way I had integrated H5_HAVE_SIGLONGJMP previously, we conditionally removed far more functionality than we needed.  As a result, Windows type alignment was lost.  This commit localizes the ifdef so it only surrounds the actual call to siglongjmp.  This is equivalent to what we've done in the past-- simply comment out the line.

Tested:
VNET on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14067 dab4d1a6-ed17-0410-a064-d1ae371a2980