Purpose: Fix Windows libtest_cstubdll output directory
Description:
Previously, the libtest_cstubdll project was creating the DLL in the project directory, rather than the Output directory.  This was a simple typo, and not caught previously because past builds had the DLL installed.  Cleaning up our batch files will make this easier to catch in the future.

Tested:
VS2005 on WinXP x64

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14099 dab4d1a6-ed17-0410-a064-d1ae371a2980