Purpose: First step in removing VS6 support in final HDF5 1.8
Description:
This is the first step in a series of commits that will remove VS6 support.  In doing so, we'll need to convert our projects to VS.NET format, to maintain VS.NET C support.  In this commit, I've converted all VS6 projects, and removed any Fortran remnants.  I haven't updated the MANIFEST because I have a few more add/deletes, but I'll do that soon.

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14427 dab4d1a6-ed17-0410-a064-d1ae371a2980