Modified metadata cache to adapt more quickly to the insertion of a large entry, or to a large increases in the size of an existing entry. This required some additions to the cache configuration structure, and thus will require changes in the metadata cache documentation.
The basic idea is to monitor the size of entries as they are loaded,
inserted, or increased in size.  If the size of the entry (or increase)
exceeds some user selected fraction of the size of the cache, increase
the size of the cache.

Note that this fix was designed quickly -- while it deals with the 
use case that exposed the problem, we may have to revisit the issue
later.

Tested serial and parallel on Phoenix, and h5committest.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14442 dab4d1a6-ed17-0410-a064-d1ae371a2980