Ported flash cache size increase code into the journaling branch -- note that both the H5C and H5C2 code have been updated.
Also checked in code to track journaling status in the super block.
Note that this code has not been tested -- but as best I can tell,
it does not break the existing regression tests.

Tested serial (debug and production) on Phoenix.  Also tested parallel
on kagiso.

Note that regression test fails on kagiso (but not on phoenix) if 
the cache2 serial tests are configured to use the core file driver.
Thus this code is check in with the core file driver optimization
of the cache2 tests disabled.  To turn it on, set the USE_CORE_DRIVER
#define to TRUE.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14594 dab4d1a6-ed17-0410-a064-d1ae371a2980