Purpose: Patch the latest Cygwin's incompatibility with timzezone variable
Description:
We discovered that there was a problem with the latest cygwin DLL, in that it doesn't recognize the global variable 'timezone'.  This is a problem on cygwin's side, not ours.  Instead, we avoid the problem in configure with a case statement.

We didn't get this out in time for the release, but we can offer a patch instead.

Tested:
Cygwin on WinXP
All changes are in a cygwin 'case' clause, so shouldn't affect other platforms

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14595 dab4d1a6-ed17-0410-a064-d1ae371a2980