Description:     Bring revision 14860 back to 1.8 branch, change log for rev 14860 is:
        Omnibus raw data I/O revisions, with wide-ranging changes and
refactoring, in order to prepare for implementing "fast append" feature.

        These changes remove the majority of the code duplication for raw data
I/O which has crept in over the last ten years and introduces a more object-
oriented design for operating on different types of dataset storage.

Description:
        Omnibus raw data I/O revisions, with wide-ranging changes and
refactoring, in order to prepare for implementing "fast append" feature.

        These changes remove the majority of the code duplication for raw data
I/O which has crept in over the last ten years and introduces a more object-
oriented design for operating on different types of dataset storage.

        Chunked storage no longer has it's own I/O routines, it is now handled
as either contiguous (if chunk is not pulled into the cache) or compact (if the
chunk is cached in memory).

        No bug or feature changes, at least intentionally... :-)

Tested on:
    Mac OS X/32 10.5.2 (amazon) w/production




git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14861 dab4d1a6-ed17-0410-a064-d1ae371a2980