Purpose: Fix bug in H5G_get_name_by_addr where buffer size is less than size of name
Description:
Backport of the bug found earlier today in HDF5 1.9.  The H5G_get_name_by_addr function makes a call to strncat writes an extra null byte when the buffer is smaller than the dataset name.  Also, comment out 2 lines that are no longer necessary because of the fix.

Tested:
kagiso


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@14984 dab4d1a6-ed17-0410-a064-d1ae371a2980