Purpose: Ignore Windows POSIX-Depreciation warnings in Visual Studio 2005
Description:
On Windows, we receive many compiler warnings because Microsoft has implemented many "security-enhanced" versions of POSIX functions.  However, the old versions still exist, so the warnings can be ignored.  By setting a preprocessor definition, we can disable many of these warnings.  On a typical HDF5 build, this reduces the number of warnings from ~1500 to ~600

Tested:
VS2005 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15120 dab4d1a6-ed17-0410-a064-d1ae371a2980